Add Cosmo Router.md
Adds a new section to the Readme (Router/Gateway) to list Gateways/Routers supporting this specification.

A router will accept client requests following the spec and will then add the files to subgraph requests following the spec again. As a result, files are being piped from client to subgraph through the router.